Move `make_content_disposition` into `HTTPResponse` module
modified:   src/OFS/ObjectManager.py
modified:   src/ZPublisher/HTTPResponse.py
deleted:    src/ZPublisher/http_header_utils.py
modified:   src/ZPublisher/tests/testHTTPResponse.py
deleted:    src/ZPublisher/tests/test_http_header_utils.py
